Allowed a user of the database to control opening / closing of the DB.
Ran some performance tests and opening and closing the databse in the
static methods was the biggest bottleneck. Allowing the user to
choose to open the db, perform many operations, and close the db both
gives that user a bit more responsibility and eliminates that bottle-
neck. I have left the static methods so that anyone who wants to
modify this without understanding when/how to open or close the db
can still interact with the database.

There were some minor changes to the record itself, I discovered the
SignalStrength class and it seems like a much better metric to use
than rssi. rssi was only intended to be temporary anyway, it was
just a placeholder until we could find better metrics.