Working on turning on the application after it is installed. Additionally, working on how to combine the data from the cellular chip and the location chips.
There are multiple difficult issues here. For example, the Android OS
has no way of informing an app that it was just installed and allowing
that app to react to the install (for example, by registering listeners
for the phone's location). I am trying to come up with a good way to
emulate this behaviour

Additionally, combining data from the various sensors can be difficult.
Not all sensors allow you to request the last known data, some of them
only allow you to listen for the data as it is generated. This can be
difficult because you can vastly waste system resources receiving data
you don't care about and just diregard.

Additionally, it seems important to remember that gps and the cell chip
are different sensors, and it seems like we may need to use two different
strategies to combine the gps and cellular data versus the network and
cellular data. This problem is made stranger when you think that the
cellular location data and the cellular signal strength data are likely
coming from the same IC.