Broke the singal CellularRecorder into two *Correlation classes, to better facilitate different handling of the association of the cellular signal to the gps versus the cellular signal to the network
Replaced the LocationListener aspect of the CellularRecorder (now the
GpsCorrelator) with a GpsStatus.Listener interface. The provider
enabled / disabled calls do not correlate to the GPS system being
turned on or off, a gps provider can be enabled but the GPS chip
can be idle. Therefore, the GpsStatus.Listener provides a much better
interface for detecting when the gps chip is on and for starting to
poll the getLastKnownLocation() method. Also, registering a
GpsStatus.Listener does not result in the gps icon in the status
bar showing up, while requesting location updates from a provider
does. Many users like to keep this icon off as much as possible to
preserve battery life. Essentially, listening on the GpsStatus and
reacting to that allows us to roughly copy the behavior of the new
'passive provider' in Android 2.2