Better module importing (relative when possible) and documentation. Moved some files into gheat from gheatae
It seems like the best method to do the importing is relative to the current
file. For example, a file inside of the gheat directory should never have to
import anything from the gheatae directory. This allows all files to have
their testing suites embedded within the file, and the user can navigate to
the directory of that file, issue a 'python filename.py' and have the tests
embedded within that file execute successfully. Later on the tests can be
moved to a different directory in order to avoid loading the tests if we do
not need to.

Also, these relative imports help to outline what is part of gheat verus
what is part of gheatae. If you are working on a file in gheat, and you
have to import something in gheatae to get your code to work, then you
likely either need to update gheat to add in an api for what you are trying
to do, or that file belongs as part of the gheat core.

Minor documentation improvements.