Made tile.py be testable from the command line. Separated gheat and gheatae better
Made the tile use the default provider and cache if none were provided. Added warning messages for using these. Also added a base point class and worked on
making all of the points implement from that. This needs some more work, as I ama bit confused wehre the geo/ code starts and ends, currently it's a bit mangledin with the gheat code but it also includes gheatae stuff